requirements.txt
langchain_core
langchain_openai
langchain_community
streamlit 
python-dotenv